+++ app/src/main/java/com/example/stock_market/Model/Operation_model.kt
@@ -1,5 +1,5 @@
 package com.example.stock_market.Model
 
-class Operation_model (var id:Int,var name:String,var type:String, var qnt:Int, var price:Float, var corretage_price:Float,var taxas:Float,var valorCompra:Float,var valorNegociacao:Float) {
+class Operation_model (var id:Int,var name:String,var type:String, var qnt:Int, var price:Double, var corretage_price:Double,var taxas:Double,var valorCompra:Double,var valorNegociacao:Double) {
 
 }
